refactor(cli): rename config module to settings
Align the CLI's settings-module filename with the MCP package: both now
expose `settings.py` holding the pure-reader models plus the
`resolve_*_settings` boundary resolver. Pure rename; imports and test
module names follow, no behavior change.
